New log storage mechanism with fixed-size partitions
- partition_t is now always created as a max-size file, and during log
initialization, a new partition is always created rather than trying to
find the end of the last partition, truncate it, and initialize (i.e.,
prime) the log buffer
- WARNING: with this commit, we really lose hope of supporting actual
system failures, since we rely on the log contents always being
consistent. This is OK for now since all our experiments perform a clean
shutdown, even when simulating a system failure.